Now uses 'jsclass'
