+++ package.json
@@ -29,6 +29,7 @@
     "express-session": "^1.11.3",
     "extend": "^2.0.1",
     "glob": "^5.0.12",
+    "jsclass": "^4.0.5",
     "morgan": "^1.6.0",
     "passport": "^0.2.2",
     "passport-bnet": "^1.1.1",
